Remove nextDirection variable from ghosts
This variable is important only for pacman, which
changes its direction even in the middle of a tile
and waits for the next intersection to perform the
direction change.
The ghosts don't need it, because they change their
direction only when they're on tile, the the direction
change happens immediately.